Domain_Finding/requirements.txt
python-dotenv
requests
openai
streamlit 
tldextract